telegram/requirements.txt
aiogram>=3.3.0
httpx>=0.27.0
structlog>=24.0.0
python-dotenv>=1.0.0
